+++ design-system/reduced-ui-pack/bin/buildSprite.js
@@ -1,6 +1,6 @@
 const fs = require('fs');
 const path = require('path');
-const glob = require('glob'); // eslint-disable-line import/no-extraneous-dependencies
+const { globSync } = require('glob'); // eslint-disable-line import/no-extraneous-dependencies
 const SVGSpriter = require('svg-sprite'); // eslint-disable-line import/no-extraneous-dependencies
 const mkdirp = require('mkdirp'); // eslint-disable-line import/no-extraneous-dependencies
 const bolt = require('bolt');
@@ -35,14 +35,14 @@ bolt.getWorkspaces({ only: '@atlaskit/icon' }).then(([pkg]) => {
   const spriter = new SVGSpriter(spriterConfig);
 
   // Add SVG source files from 'ak-icon'
-  glob
-    .sync(path.join(iconPackageDir, 'svgs/**/*.svg'), {})
-    .forEach((svgFile) => {
+  globSync(path.join(iconPackageDir, 'svgs/**/*.svg'), {}).forEach(
+    (svgFile) => {
       // eslint-disable-next-line no-console
       console.log(svgFile);
       const svgContents = fs.readFileSync(svgFile, { encoding: 'utf-8' });
       spriter.add(svgFile, path.basename(svgFile), svgContents);
-    });
+    },
+  );
 
   // Compile the sprite
   spriter.compile((error, result) => {