+++ editor/editor-plugin-editor-viewmode/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/editor-plugin-editor-viewmode
 
+## 1.0.9
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 1.0.8
 
 ### Patch Changes